refactor: menu-group 패키지 이동
