+++ src/test/java/kitchenpos/Fixtures.java
@@ -9,6 +9,7 @@ import java.util.Random;
 import java.util.UUID;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.product.domain.Product;
 
 public class Fixtures {
